fixed dkms
Previous version of the Makefile has been using $(uname -r) for
getting info about kernel version. But this command returns version
of current running kernel and not the version the module is compiled
for. This change should fix problems with automatic installation
of kernel modules after kernel has been upgraded.